Pass timer id as first argument to timer callbacks (BREAKING ABI)
This is a breaking ABI change in the timer callback API.  The reason is
to ensure even timer callbacks can call pev_timer_set() without having
to use global variables or waste the void *arg.

To allow callbacks to read their current timeout/period, as used in the
example.c code, we introduce a new member to struct pev, gettime, which
is only set while the callback is active.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>